# custom/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SuiteCRM COBOL Service API",
    "description": "REST API for COBOL-powered banking features in SuiteCRM",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@suitecrm-cobol.com"
    }
  },
  "servers": [
    {
      "url": "https://your-instance.com/api/v1/cobol",
      "description": "Production server"
    },
    {
      "url": "https://sandbox.your-instance.com/api/v1/cobol",
      "description": "Sandbox server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
  "paths": {
    "/calculate": {
      "post": {
        "summary": "Execute Financial Calculation",
        "description": "Perform COBOL-powered financial calculations with banking precision",
        "operationId": "calculate",
        "tags": ["Calculations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calculation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "CalculationRequest": {
        "type": "object",
        "required": ["type", "parameters"],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "LOAN-PAYMENT",
              "COMPOUND-INTEREST",
              "AMORTIZATION",
              "ROI-CALCULATOR",
              "MORTGAGE-CALCULATOR",
              "CURRENCY-CONVERSION",
              "RISK-ASSESSMENT",
              "TAX-CALCULATION"
            ]
          },
          "parameters": {
            "type": "object",
            "properties": {
              "principal": {
                "type": "number",
                "format": "double"
              },
              "rate": {
                "type": "number",
                "format": "double"
              },
              "term_years": {
                "type": "integer"
              },
              "frequency": {
                "type": "string",
                "enum": ["monthly", "quarterly", "annually"]
              }
            }
          },
          "currency": {
            "type": "string",
            "default": "USD"
          }
        }
      },
      "CalculationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "calculation_id": {
            "type": "string"
          },
          "result": {
            "type": "object"
          },
          "execution_time": {
            "type": "number",
            "format": "double"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          }
        }
      }
    }
  }
}